reorganized plots and tables
